% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XinyuJiangToolsRfunctions.R
\name{func10}
\alias{func10}
\title{Quiz 2 - 3}
\usage{
func10(a)
}
\arguments{
\item{a}{matrix}
}
\value{
matrix
}
\description{
a function that takes a numeric matrix and standardizes its columns
}
