% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XinyuJiangToolsRfunctions.R
\name{func11}
\alias{func11}
\title{HW 2 - 1}
\usage{
func11(a)
}
\arguments{
\item{a}{matrix}
}
\value{
object
}
\description{
a function that takes a numeric matrix and standardizes its columns. but do it without loops.
}
