% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XinyuJiangToolsRfunctions.R
\name{func12}
\alias{func12}
\title{HW 2 - 2}
\usage{
func12(X, MARGIN, FUN, ...)
}
\arguments{
\item{X}{matrix}

\item{MARGIN}{margin of object to work}

\item{FUN}{function}

\item{...}{probability}
}
\value{
array
}
\description{
a function just like the function array in the R base package,
}
