% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XinyuJiangToolsRfunctions.R
\name{\%func9\%}
\alias{\%func9\%}
\title{Quiz 2 - 2}
\usage{
a \%func9\% x
}
\arguments{
\item{a}{matrix}

\item{x}{vector}
}
\value{
object
}
\description{
calculates $x^T A^{-1} x$ but it is a binary operator rather than an apparent function call,
}
